Add descriptive PHPDoc to skills classes and refactor sync methods
